feat: :wrench: 自分の投稿のみ操作できるようにする

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Service\TweetService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -13,11 +15,15 @@ class DeleteController extends Controller
      * Handle the incoming request.
      *
      * @param Request $request
+     * @param TweetService $tweetService
      * @return RedirectResponse
      */
-    public function __invoke(Request $request): RedirectResponse
+    public function __invoke(Request $request, TweetService $tweetService): RedirectResponse
     {
         $tweetId = (int)$request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
